Fixed names. change col to col-lg
