+++ .gitignore
@@ -8,7 +8,6 @@
 
 # production
 /build
-
 # misc
 .env
 .DS_Store
