+++ src/Homes/Content.js
@@ -5,11 +5,9 @@ import GoogleMap from "./GoogleMap";
 
 export default function() {
   return (
-    <div>
-      <div className="container">
-        <List />
-        <GoogleMap center={{ lat: 19.23, lng: 75.62 }} zoom={8} />
-      </div>
+    <div className="container">
+      <List />
+      <GoogleMap center={{ lat: 19.23, lng: 75.62 }} zoom={8} />
     </div>
   );
 }
