+++ src/Homes/data.js
@@ -12,41 +12,41 @@ export default [
     price: 82,
     name: "La Salentina, see, nature & relax",
     roomtype: "Entire house  ·  9 beds",
-    superhot: "97 · Superhost"
+    superhost: "97 · Superhost"
   },
   {
     img: imgCard2,
     price: 82,
     name: "Your private 3 bedr. riad and exclusi…",
     roomtype: "Entrie house  ·  5 beds",
-    superhot: "161 · Superhost"
+    superhost: "161 · Superhost"
   },
   {
     img: imgCard3,
     price: 200,
     name: "Dreamy Tropical Tree House",
     roomtype: "Entrie treehouse  ·  1 bed",
-    superhot: "364 · Superhost"
+    superhost: "364 · Superhost"
   },
   {
     img: imgCard4,
     price: 110,
     name: "Best location old town luxury loft",
     roomtype: "Entrie apartment  ·  1 bed",
-    superhot: "369 · Superhost"
+    superhost: "369 · Superhost"
   },
   {
     img: imgCard5,
     price: 83,
     name: "Lussuoso. Vista incantevole.",
     roomtype: "Entrie apartment  ·  6 bed",
-    superhot: "105 · Superhost"
+    superhost: "105 · Superhost"
   },
   {
     img: imgCard6,
     price: 72,
     name: "In the historical center of Lecce",
     roomtype: "Entire house  ·  3 beds",
-    superhot: "221 · Superhost "
+    superhost: "221 · Superhost "
   }
 ];
