+++ src/Rooms/index.js
@@ -1,0 +1,13 @@
+import React from "react";
+import styled from "styled-components";
+
+const PaddingHeader = styled.div`height: 81px;`;
+
+export default function() {
+  return (
+    <div>
+      <PaddingHeader />
+      <h1>Hello</h1>
+    </div>
+  );
+}
