+++ src/UI/TitleSectionAndAllSee.js
@@ -53,8 +53,8 @@ width: 235px;
 
 export default function(props) {
   return (
-    <Row className="col-12 col-sm-12 col-xs-12">
-      <div className="col-9 col-sm-9 col-xs-9">
+    <Row className="col-lg-12 col-sm-12 col-xs-12">
+      <div className="col-lg-9 col-sm-9 col-xs-9">
         <Name>{props.name}</Name>
       </div>
       <Wrapper to={props.link}>
